Moved client interceptor to the interceptor package
