+++ src/main/java/io/temporal/client/WorkflowClientOptions.java
@@ -21,6 +21,7 @@ package io.temporal.client;
 
 import io.temporal.common.context.ContextPropagator;
 import io.temporal.common.converter.DataConverter;
+import io.temporal.common.interceptors.WorkflowClientInterceptor;
 import io.temporal.enums.v1.QueryRejectCondition;
 import java.lang.management.ManagementFactory;
 import java.util.Arrays;
