+++ src/main/java/io/temporal/common/interceptors/WorkflowClientInterceptorBase.java
@@ -17,11 +17,15 @@
  *  permissions and limitations under the License.
  */
 
-package io.temporal.client;
+package io.temporal.common.interceptors;
 
+import io.temporal.client.ActivityCompletionClient;
+import io.temporal.client.WorkflowOptions;
+import io.temporal.client.WorkflowStub;
 import io.temporal.common.v1.WorkflowExecution;
 import java.util.Optional;
 
+/** Convenience base class for WorkflowClientInterceptor implementations. */
 public class WorkflowClientInterceptorBase implements WorkflowClientInterceptor {
 
   @Override
