+++ src/main/java/io/temporal/internal/sync/TestWorkflowEnvironmentInternal.java
@@ -22,10 +22,10 @@ package io.temporal.internal.sync;
 import com.google.common.collect.ObjectArrays;
 import io.temporal.client.ActivityCompletionClient;
 import io.temporal.client.WorkflowClient;
-import io.temporal.client.WorkflowClientInterceptor;
 import io.temporal.client.WorkflowClientOptions;
 import io.temporal.client.WorkflowOptions;
 import io.temporal.client.WorkflowStub;
+import io.temporal.common.interceptors.WorkflowClientInterceptor;
 import io.temporal.common.v1.WorkflowExecution;
 import io.temporal.internal.testservice.TestWorkflowService;
 import io.temporal.serviceclient.WorkflowServiceStubs;
