+++ src/main/java/io/temporal/internal/sync/WorkflowClientInternal.java
@@ -25,11 +25,11 @@ import com.uber.m3.tally.Scope;
 import io.temporal.client.ActivityCompletionClient;
 import io.temporal.client.BatchRequest;
 import io.temporal.client.WorkflowClient;
-import io.temporal.client.WorkflowClientInterceptor;
 import io.temporal.client.WorkflowClientOptions;
 import io.temporal.client.WorkflowOptions;
 import io.temporal.client.WorkflowStub;
 import io.temporal.common.converter.DataConverter;
+import io.temporal.common.interceptors.WorkflowClientInterceptor;
 import io.temporal.common.v1.WorkflowExecution;
 import io.temporal.internal.external.GenericWorkflowClientExternalImpl;
 import io.temporal.internal.external.ManualActivityCompletionClientFactory;
